Add PER matrix
Previouly, PER (packet error rate) was calculated by get_error_prob().
This patch implements a option to determine PER with outer matrix
defined in IEEE 802.11ax. The matrix is defined in Appendix 3 of
"11-14-0571-12 TGax Evaluation Methodology".
https://mentor.ieee.org/802.11/dcn/14/11-14-0571-12-00ax-evaluation-methodology.docx
There is no data for 1, 2, 5.5, 11 and 9Mbps, so we made data for them
by shifted BPSK1/2 datum.

Signed-off-by: Masashi Honma <masashi.honma@gmail.com>