drop built-in PER matrix
I had asked for this, but after experimenting a bit I found that
the curves are not very well specified around the transitions which
makes it hard to use this tool for testing rate controllers and
the like.  Restore the analytical PER model.  If we find that the
numbers are too far off from reality, we can consider tweaking the
functions somewhat.

Signed-off-by: Bob Copeland <me@bobcopeland.com>